+++ src/index.ts
@@ -3,3 +3,5 @@ import { onlyExportComponents } from "./only-export-components.ts";
 export const rules = {
   "only-export-components": onlyExportComponents,
 };
+// eslint-disable-next-line @arnaud-barre/no-default-export
+export default { rules };