+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## Unreleased
+## 4.0.12
 
+- Support type assertion on default export (fixes #48)
 - Add default export to fix usage with jiti (fixes #50)
 
 ## 0.4.11